Development

Closes #64

See merge request RP-17-PaDReP/PyPaDRe!15

+++ padre/backend/file.py
@@ -10,32 +10,32 @@ import re
 import shutil
 import uuid
 
+from deprecated import deprecated
+
 from padre.backend.serialiser import JSonSerializer, PickleSerializer
-from padre.base import result_logger
-from padre.datasets import Dataset, Attribute
-from padre.experiment import Experiment
+from padre.core.datasets import Dataset, Attribute
+from padre.core import Experiment
 
 
-def _get_path(root_dir, name):
+def _get_path(root_dir, name, create=True):
     # internal get or create path function
     _dir = os.path.expanduser(os.path.join(root_dir, name))
-    if not os.path.exists(_dir):
+    if not os.path.exists(_dir) and create:
         os.mkdir(_dir)
     return _dir
 
 
-def _dir_list(root_dir, search_id, search_metadata, strip_postfix=None):
-    # todo implement search in metadata using some kind of syntax (e.g. jsonpath, grep),
-    # then search the metadata files one by one.
-    files = [f for f in os.listdir(root_dir) if strip_postfix==None or f.endswith(strip_postfix)]
-    if search_id is not None:
-        rid = re.compile(search_id)
+def _dir_list(root_dir, matcher, strip_postfix=""):
+    files = [f for f in os.listdir(root_dir) if f.endswith(strip_postfix)]
+    if matcher is not None:
+        rid = re.compile(matcher)
         files = [f for f in files if rid.match(f)]
 
-    if search_metadata is not None:
-        raise NotImplemented()
-
-    return [file[:-1*len(strip_postfix)] for file in files if file is not None and len(file)>=len(strip_postfix)]
+    if len(strip_postfix) == 0:
+        return files
+    else:
+        return [file[:-1*len(strip_postfix)] for file in files
+            if file is not None and len(file) >= len(strip_postfix)]
 
 
 class PadreFileBackend(object):
@@ -53,6 +53,7 @@ class PadreFileBackend(object):
         self._dataset_repository = DatasetFileRepository(os.path.join(root_dir, "datasets"))
         self._experiment_repository = ExperimentFileRepository(os.path.join(root_dir, "experiments"),
                                                                self._dataset_repository)
+
 
     @property
     def datasets(self):
@@ -90,12 +91,24 @@ class ExperimentFileRepository:
     So logically they would belong to the splits.
     However, for convenience reasons they are aggregated at the run level.
     """
+    _file = None
 
     def __init__(self, root_dir, data_repository):
         self.root_dir = _get_path(root_dir, "")
+        self._split_dir = _get_path(root_dir, "")
         self._metadata_serializer = JSonSerializer
         self._binary_serializer = PickleSerializer
         self._data_repository = data_repository
+        self._file = None
+
+    def __del__(self):
+        """
+        Destructor that closes the opened log file at the end of the experiments
+        :return:
+        """
+        if self._file is not None:
+            self._file.close()
+            self._file = None
 
     def _dir(self, ex_id, run_id=None, split_num=None):
         r = [str(ex_id)+".ex"]
@@ -105,48 +118,68 @@ class ExperimentFileRepository:
                 r.append(str(split_num)+".split")
         return r
 
-    def list_experiments(self, search_id=None, search_metadata=None):
+    def list_experiments(self, search_id=".*", search_metadata=None):
         """
         list the experiments available
         :param search_id:
         :param search_metadata:
         :return:
         """
+        returns = _dir_list(self.root_dir, search_id+".ex", ".ex")
+        # todo: implement search metadata filter
+        return returns
 
-        return _dir_list(self.root_dir, search_id, search_metadata, ".ex")
+    def delete_experiments(self, search_id=".*", search_metadata=None):
+        """
+        list the experiments available
+        :param search_id:
+        :param search_metadata:
+        :return:
+        """
+        dirs = _dir_list(self.root_dir, search_id+".ex")
+        # todo: implement search metadata filter
+        for d in dirs:
+            shutil.rmtree(_get_path(self.root_dir, d, False))
 
-    def put_experiment(self, experiment, allow_overwrite=False):
+
+    def put_experiment(self, experiment, append_runs=False, allow_overwrite=True):
         """
         Stores an experiment to the file. Only metadata, hyperparameter and the workflow is stored.
         :param experiment:
-        :param allow_overwrite: True if an existing experiment can be overwritten
+        :param append_runs: True if runs can be added to an existing experiment.
+        If false, any existing experiment will be removed
         :return:
         """
+
         if experiment.id is None:  #  this is a new experiment
             if experiment.name is None or experiment.name == "":
-                experiment.id = uuid.uuid1()
+                experiment.id = uuid.uuid4()
             else:
                 experiment.id = experiment.name
         dir = os.path.join(self.root_dir, *self._dir(experiment.id))
         if os.path.exists(dir) and not allow_overwrite:
             raise ValueError("Experiment %s already exists." +
                              "Overwriting not explicitly allowed. Set allow_overwrite=True")
+
+        # Create the log file for the experiment here
+        if self._file is not None:
+            self._file.close()
+            self._file = None
+
         if os.path.exists(dir):
-            shutil.rmtree(dir)
-        os.mkdir(dir)
+            if not append_runs:
+                shutil.rmtree(dir)
+                os.mkdir(dir)
+        else:
+            os.mkdir(dir)
+
+        self._file = open(os.path.join(dir, "log.txt"), "a")
+
         with open(os.path.join(dir, "metadata.json"), 'w') as f:
             f.write(self._metadata_serializer.serialise(experiment.metadata))
-        with open(os.path.join(dir, "hyperparameter.json"), 'w') as f:
-            params = experiment.hyperparameters()
-            for key in params:
-                # This writes all data present within the params to the JSON file
-                f.write(self._metadata_serializer.serialise((params[key])))
-
-                # This is to write only the hyper parameters of each run to the JSON file
-                #f.write(self._metadata_serializer.serialise((params[key]['hyper_parameters'])))
 
         with open(os.path.join(dir, "workflow.bin"), 'wb') as f:
-            f.write(self._binary_serializer.serialise(experiment._workflow))
+            f.write(self._binary_serializer.serialise(experiment.workflow))
 
     def get_experiment(self, id_, load_workflow=True):
         dir = os.path.join(self.root_dir, *self._dir(id_))
@@ -175,7 +208,20 @@ class ExperimentFileRepository:
         :param search_metadata:
         :return:
         """
-        return _dir_list(os.path.join(self.root_dir, *self._dir(experiment_id)), search_id, search_metadata, ".run")
+        # todo: impelement filter on metadata
+        return _dir_list(os.path.join(self.root_dir, *self._dir(experiment_id)), search_id, ".run")
+
+    def put_experiment_configuration(self, experiment):
+        """
+        Writes the experiment details as a json file
+
+        :param experiment: the experiment to be written to the disk
+
+        :return:
+        """
+        if experiment.experiment_configuration is not None:
+            with open(os.path.join(self.root_dir, *self._dir(experiment.id), "experiment.json"), 'w') as f:
+                f.write(self._metadata_serializer.serialise(experiment.experiment_configuration))
 
     def put_run(self, experiment, run):
         """
@@ -185,27 +231,26 @@ class ExperimentFileRepository:
         :return:
         """
         if run.id is None:  # this is a new experiment
-            run.id = uuid.uuid1()
+            run.id = uuid.uuid4()
 
         dir = os.path.join(self.root_dir, *self._dir(experiment.id, run.id))
         if os.path.exists(dir):
             shutil.rmtree(dir)
         os.mkdir(dir)
 
-        # Set the directory for logging
-        result_logger.set_log_directory(os.path.join(self.root_dir, *self._dir(experiment.id, run.id)))
-
         with open(os.path.join(dir, "metadata.json"), 'w') as f:
             f.write(self._metadata_serializer.serialise(experiment.metadata))
 
+        #Commented for pytorch integration
+
         with open(os.path.join(dir, "hyperparameter.json"), 'w') as f:
             params = experiment.hyperparameters()
-            for key in params:
-                # This writes all data present within the params to the JSON file
-                f.write(self._metadata_serializer.serialise((params[key])))
+            # This writes all data present within the params to the JSON file
+            f.write(self._metadata_serializer.serialise(params))
+
 
         with open(os.path.join(dir, "workflow.bin"), 'wb') as f:
-            f.write(self._binary_serializer.serialise(experiment._workflow))
+            f.write(self._binary_serializer.serialise(experiment.workflow))
 
     def get_run(self, ex_id, run_id):
         """
@@ -240,14 +285,17 @@ class ExperimentFileRepository:
         :return:
         """
         if split.id is None:  # this is a new experiment
-            split.id = str(split.number)+":"+str(uuid.uuid1())
+            split.id = uuid.uuid4()
 
-        dir = os.path.join(self.root_dir, *self._dir(experiment.id, run.id, split.id))
+        split_id = str(split.number) + "_" + str(split.id)
+
+        dir = os.path.join(self.root_dir, *self._dir(experiment.id, run.id, split_id))
         if os.path.exists(dir):
             shutil.rmtree(dir)
         os.mkdir(dir)
         with open(os.path.join(dir, "metadata.json"), 'w') as f:
             f.write(self._metadata_serializer.serialise(experiment.metadata))
+        self._split_dir = dir
 
     def get_split(self, ex_id, run_id, split_id):
         """
@@ -256,15 +304,140 @@ class ExperimentFileRepository:
         :param run_id:
         :return:
         """
-        dir = os.path.join(self.root_dir, *self._dir(ex_id, run_id, split_id))
+        dir_ = os.path.join(self.root_dir, *self._dir(ex_id, run_id, split_id))
 
-        with open(os.path.join(dir, "metadata.json"), 'r') as f:
+        with open(os.path.join(dir_, "metadata.json"), 'r') as f:
             metadata = self._metadata_serializer.deserialize(f.read())
 
         return None
 
+    def put_results(self, experiment, run, split, results):
+        """
+        Write the results of a split to the backend
+
+        :param experiment: Experiment ID
+        :param run_id: Run ID of the current experiment run
+        :param split_id: Split id
+        :param results: results to be written to the backend
+
+        :return: None
+        """
+        split_id = str(split.number) + "_" + str(split.id)
+        dir_ = os.path.join(self.root_dir, *self._dir(experiment.id, run.id, split_id))
+        with open(os.path.join(dir_, "results.json"), 'w') as f:
+            f.write(self._metadata_serializer.serialise(results))
+
+    def put_metrics(self, experiment, run, split, metrics):
+        """
+        Writes the metrics of a split to the backend
+
+        :param experiment: Experiment ID
+        :param run: Run Id of the experiment
+        :param split: Split ID
+        :param metrics: dictionary containing all the required metrics to be written to the backend
+
+        :return: None
+        """
+        split_id = str(split.number) + "_" + str(split.id)
+        dir_ = os.path.join(self.root_dir, *self._dir(experiment.id, run.id, split_id))
+        with open(os.path.join(dir_, "metrics.json"), 'w') as f:
+            f.write(self._metadata_serializer.serialise(metrics))
+
     def _do_print(self):
         return True
+
+    def log(self, message):
+        """
+        This function logs all the messages to a file backend
+
+        :param message: Message to be written to a file
+
+        :return:
+        """
+        if self._file is None:
+            self._file = open(os.path.join(self.root_dir, "log.txt"), "a")
+
+        self._file.write(message + "\n")
+
+    def log_experiment_progress(self, curr_value, limit, phase):
+        """
+
+        :param curr_value:
+        :param limit:
+        :param phase:
+        :return:
+        """
+        if self._file is None:
+            self._file = open(os.path.join(self.root_dir, "log.txt"), "a")
+
+        self._file.write("EXPERIMENT PROGRESS: {curr_value}/{limit}. phase={phase} \n".format(phase=phase,
+                                                                                              curr_value=curr_value,
+                                                                                              limit=limit))
+
+    def log_run_progress(self, curr_value, limit, phase):
+        """
+
+        :param curr_value:
+        :param limit:
+        :param phase:
+        :return:
+        """
+        if self._file is None:
+            self._file = open(os.path.join(self.root_dir, "log.txt"), "a")
+
+        self._file.write("RUN PROGRESS: {curr_value}/{limit}. phase={phase} \n".format(phase=phase,
+                                                                                       curr_value=curr_value,
+                                                                                       limit=limit))
+
+    def log_split_progress(self, curr_value, limit, phase):
+        """
+
+        :param curr_value:
+        :param limit:
+        :param phase:
+        :return:
+        """
+        if self._file is None:
+            self._file = open(os.path.join(self.root_dir, "log.txt"), "a")
+
+        self._file.write("SPLIT PROGRESS: {curr_value}/{limit}. phase={phase} \n".format(phase=phase,
+                                                                                         curr_value=curr_value,
+                                                                                         limit=limit))
+
+    def log_progress(self, message, curr_value, limit, phase):
+        """
+
+        :param message:
+        :param curr_value:
+        :param limit:
+        :param phase:
+        :return:
+        """
+        if self._file is None:
+            self._file = open(os.path.join(self.root_dir, "log.txt"), "a")
+
+        self._file.write("PROGRESS: {curr_value}/{limit}. phase={phase}. Message:{message} \n".
+                         format(phase=phase, curr_value=curr_value, limit=limit,
+                                message=message))
+
+    def log_end_experiment(self):
+        self._file.close()
+        self._file = None
+
+    def log_model(self, model, framework, modelname, finalmodel=False):
+        """
+        Logs an intermediate model to the backend
+        :param model: Model to be logged
+        :param framework: Framework of the model
+        :param modelname: Name of the intermediate model
+        :param finalmodel: Boolean value indicating whether the model is the final one or not
+        :return:
+        """
+        if framework == 'pytorch':
+            import torch
+            import os
+            path = os.path.join(self._split_dir,modelname)
+            torch.save(model, path)
 
 
 
@@ -287,16 +460,25 @@ class DatasetFileRepository(object):
         self._metadata_serializer = JSonSerializer
         self._data_serializer = PickleSerializer
 
-    def list_datasets(self, search_id=None, search_metadata=None):
+    def list(self, search_name=None, search_metadata=None) -> list:
         """
         List all data sets in the repository
         :param search_name: regular expression based search string for the title. Default None
         :param search_metadata: dict with regular expressions per metadata key. Default None
         """
-        return _dir_list(self.root_dir, search_id, search_metadata)
+        # todo apply the search metadata filter.
+        dirs = _dir_list(self.root_dir, search_name)
+        return [self.get(dir, metadata_only=True) for dir in dirs]
 
-    def put_dataset(self, dataset):
-        _dir = _get_path(self.root_dir, dataset.id)
+
+    def put(self, dataset: Dataset)-> None:
+        """
+        stores the provided dataset into the file backend under the directory `dataset.id`
+        (file `data.bin` contains the binary and file `metadata.json` contains the metadata)
+        :param dataset: dataset to put.
+        :return:
+        """
+        _dir = _get_path(self.root_dir, str(dataset.id))
         try:
             if dataset.has_data():
                 with open(os.path.join(_dir, "data.bin"), 'wb') as f:
@@ -311,8 +493,11 @@ class DatasetFileRepository(object):
             shutil.rmtree(_dir)
             raise e
 
+    @deprecated(reason="use put")
+    def put_dataset(self, dataset):
+        self.put(dataset)
 
-    def get_dataset(self, id, metadata_only=False):
+    def get(self, id, metadata_only=False):
         """
         Fetches a data set with `name` and returns it (plus some metadata)
 
@@ -323,21 +508,35 @@ class DatasetFileRepository(object):
 
         with open(os.path.join(_dir, "metadata.json"), 'r') as f:
             metadata = self._metadata_serializer.deserialize(f.read())
-            attributes = metadata.pop("attributes")
+        attributes = metadata.pop("attributes")
+        # print(type(metadata))
+        ds = Dataset(id, **metadata)
+        #sorted(attributes, key=lambda a: a["index"])
+        #assert sum([int(a["index"]) for a in attributes]) == len(attributes) * (
+        #    len(attributes) - 1) / 2  # check attribute correctness here
+        attributes = [Attribute(a["name"], a["measurementLevel"], a["unit"], a["description"],
+                               a["defaultTargetAttribute"], a["context"], a["index"])
+                     for a in attributes]
+        ds.set_data(None,attributes)
+        if os.path.exists(os.path.join(_dir, "data.bin")):
+            def __load_data():
+                with open(os.path.join(_dir, "data.bin"), 'rb') as f:
+                    data = self._data_serializer.deserialize(f.read())
+                return data, attributes
+            ds.set_data(__load_data)
+        return ds
 
-        ds = Dataset(id, metadata)
-        sorted(attributes, key=lambda a: a["index"])
-        assert sum([int(a["index"]) for a in attributes]) == len(attributes) * (
-            len(attributes) - 1) / 2  # check attribute correctness here
-        ds.set_data(None,
-                    [Attribute(a["name"], a["measurementLevel"], a["unit"], a["description"],
-                               a["defaultTargetAttribute"])
-                     for a in attributes])
-        if metadata_only:
-            return ds
-        elif os.path.exists(os.path.join(_dir, "data.bin")):
-            data = None
-            with open(os.path.join(_dir, "data.bin"), 'rb') as f:
-                data = self._data_serializer.deserialize(f.read())
-            ds.set_data(data, ds.attributes)
+    def delete(self, id):
+        """
+        :param id:
+        :return:
+        """
+        _dir = _get_path(self.root_dir, id)
+        if os.path.exists(_dir):
+            shutil.rmtree(_dir)
+
+
+    @deprecated(reason="use get")
+    def get_dataset(self, id, metadata_only=False):
+        return self.get(id, metadata_only)
